+++ NEWS.md
@@ -14,6 +14,9 @@ NEW FEATURES :
 * `createCluster()` parameter `list_pollutants` default value to NULL.
 * `createBindingConstraint()` parameter `coefficients` must be alphabetically ordered.
 * `.createCluster()` default matrix in API mode.
+* `createClusterST()` : add a control to check if a cluster exists before running actions.
+* `editClusterST()` : add a control to check if a cluster exists before running actions.
+* `.removeCluster()` : add a control to check if a cluster exists before running actions in st-storage mode.
 
 BUGFIXES : 
 
@@ -24,7 +27,15 @@ BUGFIXES :
 * Fix `createCluster()` and `editCluster()` parameter `list_pollutants` stop if Antares Version < 8.6.0
 * `getJobs()` no longer returns duplicates and displays the two new columns `owner_id` and `owner_name`.
 * Allow the user to set symbol or full name as argument series in `updateScenarioBuilder()`
+* `scenarioBuilder()` matrix has the same row repeated if the area is not rand
 * Fix `createLink()` to update opts in API mode.
+* Fix `editClusterST()` : can not edit a cluster if it does not exist in API mode.
+* `updateScenarioBuilder()` works for NTC part : allow cartesian in the merge.
+* `api_command_execute()` :  
+  - no longer deletes a command  
+  - displays a success message for a study or variant
+
+
 
 # antaresEditObject 0.6.1